Class to test html generation

+++ src/test/scala/flaky/Unzip.scala
@@ -4,7 +4,7 @@ import java.io.File
 
 trait Unzip {
 
-  def unzip(zipped:File, unzipDir:File): Unit = {
+  def unzip(zipped: File, unzipDir: File, deleteOnExit: Boolean = true): Unit = {
     import java.io.{FileInputStream, FileOutputStream}
     import java.util.zip.ZipInputStream
     val fis = new FileInputStream(zipped)
@@ -18,10 +18,14 @@ trait Unzip {
         if (file.isDirectory) {
           val dir = new File(unzipDir, file.getName)
           dir.mkdirs()
-          dir.deleteOnExit()
+          if (deleteOnExit){
+            dir.deleteOnExit()
+          }
         } else {
           val file1 = new File(unzipDir, file.getName)
-          file1.deleteOnExit()
+          if (deleteOnExit){
+            file1.deleteOnExit()
+          }
           val fout = new FileOutputStream(file1)
           val buffer = new Array[Byte](1024)
           Stream.continually(zis.read(buffer)).takeWhile(_ != -1).foreach(fout.write(buffer, 0, _))
